requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-nested-inline==0.4.6
django-ninja==1.3.0
django-unfold==0.53.0
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
multidict==6.3.0
openpyxl==3.1.5
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
